+++ CHANGELOG.md
@@ -75,6 +75,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - custom items in the chest conversation IO are now defined in the `properties` section on an option
 - Reworked the config.yml file to be more clean and structured
     - `item.backpack.cancel_button` and `item.backpack.compass_button` do not have a `DEFAULT` value anymore, instead you need to specify an item
+- LoadDataEvent is now called before and after the data is loaded with a new enum `LoadDataEvent.State` that indicates the state
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation